+++ pegasus_fr.ts
@@ -1375,7 +1375,7 @@ dans les répertoires suivants :</translation>
     <message>
         <location filename="../src/frontend/menu/settings/InterfaceMain.qml" line="179"/>
         <source>Once enabled, you can run emulators in separate windows and keep pegasus/theme activated</source>
-        <translation>Activé, tu peux lancer l'emulateur et garder Pegasus/Theme actif </translation>
+        <translation>Activé, tu peux lancer l&apos;emulateur et garder Pegasus/Theme actif </translation>
     </message>
     <message>
         <location filename="../src/frontend/menu/settings/InterfaceMain.qml" line="191"/>
@@ -2078,7 +2078,7 @@ dans les répertoires suivants :</translation>
         <translation>choisir la sortie</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="412"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="423"/>
         <source>Enable or disable Wifi</source>
         <translation>Pour activer ou désactiver Wifi</translation>
     </message>
@@ -2087,132 +2087,142 @@ dans les répertoires suivants :</translation>
         <translation type="vanished">Systèmes</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="522"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="533"/>
         <source>System information</source>
         <translation>Information Système</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="523"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="534"/>
         <source>More information Ip, Cpu, OpenGL ...</source>
         <translation>Plus d&apos;information IP, CPU, OpenGL, etc...</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="280"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="291"/>
         <source>Storage configuration</source>
         <translation>Configuration du stockage</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="289"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="261"/>
+        <source>Video Driver</source>
+        <translation>Pilote Video</translation>
+    </message>
+    <message>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="262"/>
+        <source>Force video driver to Vulkan</source>
+        <translation>Forcer Vulkan pour le pilote video</translation>
+    </message>
+    <message>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="300"/>
         <source>Storage device</source>
         <translation>Dispositif de stockage</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="290"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="301"/>
         <source>change to over storage</source>
         <translation>changer pour un autre stockage</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="308"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="319"/>
         <source>Networks</source>
         <translation>Réseaux</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="325"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="336"/>
         <source>Ethernet Local IP :</source>
         <translation>Ethernet IP Local :</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="328"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="339"/>
         <source>Plug your cable to have network</source>
         <translation>Branchez un cable pour avoir du réseau</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="342"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="353"/>
         <source>Wifi Local IP :</source>
         <translation>Wifi IP Local :</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="342"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="353"/>
         <source>Wifi used :</source>
         <translation>Wifi utilisé :</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="345"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="356"/>
         <source>Connect your PC to any network</source>
         <translation>Connecter votre PC à un réseau</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="353"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="364"/>
         <source>Ethernet network</source>
         <translation>Réseau ethernet</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="411"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="422"/>
         <source>Wifi activation</source>
         <translation>Activation du Wifi</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="439"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="450"/>
         <source>Wifi networks</source>
         <translation>Réseaux Wifi</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="467"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="478"/>
         <source>System language</source>
         <translation>Langue du système</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="474"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="485"/>
         <source>Language</source>
         <translation>Langue</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="475"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="486"/>
         <source>Set your language interface</source>
         <translation>Définissez votre langue d&apos;interface</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="496"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="507"/>
         <source>Keyboard layout</source>
         <translation>Disposition du clavier</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="497"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="508"/>
         <source>Change keyboard layout language</source>
         <translation>Changer la langue de la disposition du clavier</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="515"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="526"/>
         <source>System</source>
         <translation>Système</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="536"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="547"/>
         <source>Debug mode</source>
         <translation>Mode débogage</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="537"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="548"/>
         <source>Give me your log baby !!! ;-)</source>
         <translation>Donne-moi ton log bébé ! !! ;-)</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="549"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="560"/>
         <source>Enable mouse support</source>
         <translation>Activer le support de la souris</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="550"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="561"/>
         <source>By default the cursor is visible if there are any pointer devices connected.</source>
         <translation>Par défaut, le curseur est visible si des dispositifs de pointage sont connectés.</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="563"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="574"/>
         <source>Enable virtual keyboard support</source>
         <translation>Activer le support du clavier virtuel</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="564"/>
+        <location filename="../src/frontend/menu/settings/SettingsMain.qml" line="575"/>
         <source>By default virtual keyboard is not activated.</source>
         <translation>Par defaut le clavier virtuel n&apos;est pas activé.</translation>
     </message>
@@ -2350,7 +2360,7 @@ dans les répertoires suivants :</translation>
     </message>
     <message>
         <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="94"/>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="214"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="216"/>
         <source>Output</source>
         <translation>Sortie</translation>
     </message>
@@ -2361,7 +2371,7 @@ dans les répertoires suivants :</translation>
     </message>
     <message>
         <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="119"/>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="242"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="244"/>
         <source>Resolution</source>
         <translation>Résolution</translation>
     </message>
@@ -2372,7 +2382,7 @@ dans les répertoires suivants :</translation>
     </message>
     <message>
         <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="145"/>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="270"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="272"/>
         <source>Frequency</source>
         <translation>Fréquences</translation>
     </message>
@@ -2392,47 +2402,47 @@ dans les répertoires suivants :</translation>
         <translation>Choisissez l&apos;orientation de votre écran principal.</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="194"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="195"/>
         <source>Secondary screen settings</source>
         <translation>Paramètres de l&apos;écran secondaire</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="215"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="217"/>
         <source>Choose your output for secondary screen.</source>
         <translation>Choisissez votre sortie pour l&apos;écran secondaire.</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="243"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="245"/>
         <source>Choose resolution for secondary screen.</source>
         <translation>Choisissez la résolution de l&apos;écran secondaire.</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="271"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="273"/>
         <source>Choose frequency for secondary screen.</source>
         <translation>Choisissez la fréquence de l&apos;écran secondaire.</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="297"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="299"/>
         <source>Rotation</source>
         <translation>Rotation</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="298"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="300"/>
         <source>Choose orientation for your secondary screen.</source>
         <translation>Choisissez l&apos;orientation de votre écran secondaire.</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="324"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="326"/>
         <source>Position</source>
         <translation>Position</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="325"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="327"/>
         <source>Choose position for your Secondary screen.</source>
         <translation>Choisissez la position de votre écran secondaire.</translation>
     </message>
     <message>
-        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="363"/>
+        <location filename="../src/frontend/menu/settings/VideoSettings.qml" line="366"/>
         <source>Apply</source>
         <translation>Appliquer</translation>
     </message>